+++ src/debugger/breakpoints/breakpoints_exception.cpp
@@ -278,7 +278,7 @@ HRESULT ExceptionBreakpoints::GetExceptionDetails(ICorDebugThread *pThread, ICor
         pDetails->evaluateName = "$exception";
 
         HRESULT Status = S_OK;
-        m_sharedEvaluator->WalkMembers(trExceptionValue, pThread, FrameLevel{0}, false, FormatSpecifier::ForceEvaluation,
+        m_sharedEvaluator->WalkMembers(trExceptionValue, pThread, FrameLevel{0}, false, FormatSpecifier::ForceEvaluation, std::string{},
             [&](ICorDebugType *, bool, const std::string &memberName,
                 const Evaluator::GetValueCallback &getValue, Evaluator::SetterData *, std::string *) -> HRESULT
             {