+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Upcoming changes compared to previous version.
 
 #### Changed
 - Updated tree-sitter version to 0.27.0.
+- Resolve array element access by name directly instead of enumerating all elements.
 
 <br>
 <br>